moves map into its own folder
