+++ src/views/components/map/map.component.js
@@ -1,6 +1,6 @@
-import { generateMap } from "../../controller";
-import { anchor } from "../../utils/clear.util";
-import createEle from "../../utils/createEle.util";
+import { generateMap } from "../../../controller";
+import { anchor } from "../../../utils/clear.util";
+import createEle from "../../../utils/createEle.util";
 import 'mapbox-gl/dist/mapbox-gl.css';
 import './map.style.scss';
